This commit speeds up the ONNX export of large models by making the following changes:

- Remove unecessary memcpy in `GetGraphProtoSize`
- In `export.cpp`, pass around a pointer to the ModelProto instead of the ModelProto itself.
- The shape inference function is very time-consuming, so only call it once when exporting (instead of twice).

Before/After of export times:
| Model  | Before | After | Speedup |
| ------------- | ------------- | -------- | ---- |
| LLaMa 7B, No constant folding  | 158s  | 74s | 2.1x |
| LLaMa 7B, With Constant folding  | 132s  | 55s | 2.4x | 
| vit_large_patch14_clip_224.openai_ft_in12k_in1k  | 11s | 6.6s  | 1.7x |
| 8 x (32K by 32K Linear + Relu)  | 72s | 28s  | 2.6x |

The shape inference function is still the slowest part of the export for these models with large weights taking up 50% or more of the export time.