@ilyasher 2023-06-12
> We should profile shape_type_inference and see why it's so slow. Maybe somewhere we are doing too many copies or too many unnecessary constant propagations.

Here's what I found while I was profiling shape inference:
1. During shape inference, each node is converted into its own separate ONNX graph to run ONNX's shape inference. Creating these ONNX nodes is expensive because it requires copying the weights.
2. The constant folding step always makes a copy of the output tensor.

As an example, here is what happens to a `torch.nn.Linear` layer when it is exported:
1. The layer is turned into Transpose -> Matmul ONNX ops. (fast)
2. The weights are copied into a new ONNX Transpose node in order to perform ONNX shape inference (slow)
3. Constant folding is run on the Transpose node and the transposed tensor is cloned and stored in ConstantValueMap (slow)
4. Now we're done with the Transpose, and move on to the Matmul.

I don't see an easy way around these copies.