@ilyasher 2023-06-27
> Publicly disabling by default the shape inference seems very tricky.

Shape inference wasn't entirely disabled, but I changed it to only run once at the end instead of once in the middle and once in the end.

In any case, I reverted that change in response to a different comment.